Refactors local variable checking to only check un-changed `AstSegment`s for function shadowing if those segments make an assignment to a symbol that changed.

Before:
```
validate@local   ---------- 51.368 ops/sec
validate@0.65.26 --------- 101.435 ops/sec
```

After: 
```
validate@local   --------- 286.331 ops/sec
validate@0.65.26 --------- 102.877 ops/sec
```