@TwitchBronBron 2024-03-14
Woah! That's some serious performance improvements!!! I'll test this out in our huge project to make sure it didn't break anything, but I'm super excited about this